Create calculator element content
